app/requirements.txt
# --- Web / API ---
fastapi==0.116.1
uvicorn[standard]==0.30.1
jinja2==3.1.4
aiofiles==23.2.1
httpx==0.28.1

# --- DB / ORM ---
SQLAlchemy==2.0.31
asyncpg==0.29.0
greenlet>=3,<4
aiosqlite>=0.19,<1     # для тестов с sqlite

# --- Pydantic / Settings / Auth / Utils ---
pydantic==2.7.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
bcrypt==4.0.1
python-multipart==0.0.9

# --- Queue ---
pika==1.3.2

# --- ML ---
transformers==4.40.1
torch==2.5.1
sentencepiece==0.2.0
sacremoses==0.1.1

# --- Tests ---
pytest==8.4.1
pytest-asyncio==1.1.0
coverage==7.10.4
